+++ src/continuous.jl
@@ -1,4 +1,4 @@
-using OrdinaryDiffEq, Requires, ForwardDiff, DiffEqCallbacks
+using OrdinaryDiffEq, Requires, ForwardDiff
 import OrdinaryDiffEq.ODEProblem
 import OrdinaryDiffEq.ODEIntegrator
 
@@ -18,54 +18,52 @@ abstract type ContinuousDynamicalSystem <: DynamicalSystem end
 * `prob::ODEProblem` : The fundamental structure used to describe
   a continuous dynamical system and also used in the
   [DifferentialEquations.jl](http://docs.juliadiffeq.org/latest/index.html)
-  ecosystem.
-  Contains the system's state, the equations of motion and optionally other
+  suite.
+  Contains the system's state (`prob.u0`), the equations of motion (`prob.f`),
+  the parameters of the equations of motion (`prob.p`) and optionally other
   information like e.g. [callbacks](http://docs.juliadiffeq.org/latest/features/callback_functions.html#Event-Handling-and-Callback-Functions-1).
 * `jacob!` (function) : The function that represents the Jacobian of the system,
-  given in the format: `jacob!(t, u, J)` which means it is in-place, with the mutated
-  argument being the last (`u` **must be** `Vector`).
-* `J::Matrix{T}` : Jacobian matrix.
-
-You can use `ds.prob.u0 .= newstate` to set a new state to the system.
+  given in the format: `jacob!(J, u, p, t)` which means given a state `u`, a
+  parameter container `p` and a time `t`, it writes the system's Jacobian in-place in
+  `J`.
+* `J::Matrix{T}` : Initialized Jacobian matrix.
 
 ## Creating a `ContinuousDS`
-The equations of motion **must be** in the form `eom!(t, u, du)`,
-which means that they are **in-place** with the mutated argument
-`du` the last one. Both `u, du` **must be** `Vector`s. You can still use matrices
-in your equations of motion though! Just change `function eom!(t, u, du)` to
-```julia
-function eom!(t, u, du)
-    um = reshape(u, a, b); dum = reshape(du, a, b)
-    # equations of motion with matrix shape of a×b
-```
+The equations of motion function **must be** in the form `eom!(du, u, p, t)`,
+as requested by DifferentialEquations.jl. They are **in-place** with the mutated
+argument `du` being first. `p` stands for the parameters of the model and `t` stands
+for the time variable (independent variable). Actually using `p` and `t` inside `eom!`
+is completely optional, however *both must be used in the definition of the function*!
+Both `u, du` **must be** `Vector`s.
 
 If you have the `eom!` function, and optionally a function for the
 Jacobian, you can use the constructor
 ```julia
-ContinuousDS(state, eom! [, jacob! [, J]]; tspan = (0.0, 100.0))
+ContinuousDS(state, eom! [, jacob! [, J]];
+             tspan = (0.0, 100.0), parameters = nothing)
 ```
-with `state` the initial condition of the system.
+with `state` the initial condition of the system. `parameters` is a keyword
+corresponding to the initial parameters of the model,
+and if not given it is assumed to be `nothing` (which of course means that the
+model does not have any parameters).
 
 If instead you already have an `ODEProblem` because you also want to take advantage
 of the callback functionality of DifferentialEquations.jl, you may use the constructor
 ```julia
 ContinuousDS(odeproblem [, jacob! [, J]])
 ```
+
 If the `jacob!` is not provided by the user, it is created automatically
-using the module [`ForwardDiff`](http://www.juliadiff.org/ForwardDiff.jl/stable/)
-(which always passes `t=0` at the `eom!`) in both cases.
-
-As mentioned in our [official documentation](https://juliadynamics.github.io/DynamicalSystems.jl/latest/system_definition#example-using-functors),
-it is preferred to use Functors for both the equations of motion and the Jacobian.
-
-To interfece *towards* DifferentialEquations.jl use `ODEIntegrator(ds, stuff...)`
-and `ODEProblem(ds, stuff...)`.
-Notice that you can have performance gains for stiff methods by
-explicitly adding a Jacobian caller for DifferentialEquations.jl by defining
-`eom!(::Type{Val{:jac}}, t, u, J) = jacob!(t, u, J)`.
+using the module [`ForwardDiff`](http://www.juliadiff.org/ForwardDiff.jl/stable/),
+which always passes `p = odeproblem.p, t=0` at the `eom!` (this interaction is
+well-behavied for all functions exported by the DynamicalSystems.jl suite because
+the parameter container is passed by reference and thus mutations in parameters
+propagate correctly).
 
 `ContinuousDS` by default are evolved using solver `Vern9()` and tolerances
 `:abstol => 1e-9, :reltol => 1e-9`.
+
+See also [`set_state!`](@ref).
 """
 struct ContinuousDS{T<:Number, ODE<:ODEProblem, JJ} <: ContinuousDynamicalSystem
     prob::ODE
@@ -75,10 +73,10 @@ struct ContinuousDS{T<:Number, ODE<:ODEProblem, JJ} <: ContinuousDynamicalSystem
     function ContinuousDS{T,ODE,JJ}(prob, j!, J) where {T,ODE,JJ}
 
         typeof(prob.u0) <: Vector || throw(ArgumentError(
-        "Currently we only support vectors as states, "*
+        "We only support vectors as states, "*
         "see the documentation string of `ContinuousDS`."
         ))
-        j!(0, prob.u0, J)
+        j!(J, prob.u0, prob.p, 0)
 
         eltype(prob.u0) == eltype(J) || throw(ArgumentError(
         "The state and the Jacobian must have same type of numbers."))
@@ -88,7 +86,10 @@ struct ContinuousDS{T<:Number, ODE<:ODEProblem, JJ} <: ContinuousDynamicalSystem
 end
 
 # Constructors with Jacobian:
-ContinuousDS(prob::ODE, j!::JJ, J::Matrix{T}) where {T<:Number, ODE<:ODEProblem, JJ} = ContinuousDS{T, ODE,JJ}(prob, j!, J)
+function ContinuousDS(prob::ODE, j!::JJ, J::Matrix{T}) where
+    {T<:Number, ODE<:ODEProblem, JJ}
+    ContinuousDS{T, ODE,JJ}(prob, j!, J)
+end
 
 function ContinuousDS(prob::ODEProblem, j!)
 
@@ -97,10 +98,11 @@ function ContinuousDS(prob::ODEProblem, j!)
 end
 
 function ContinuousDS(state, eom!, j!,
-    J = zeros(eltype(state), length(state), length(state)); tspan=(0.0, 100.0))
+    J = zeros(eltype(state), length(state), length(state));
+    tspan=(0.0, 100.0), parameters = nothing)
 
-    j!(0.0, state, J)
-    problem = ODEProblem{true}(eom!, state, tspan)
+    j!(J, state, parameters, 0)
+    problem = ODEProblem{true}(eom!, state, tspan, parameters)
 
     return ContinuousDS(problem, j!, J)
 end
@@ -108,35 +110,35 @@ end
 
 # Constructors without Jacobian:
 function ContinuousDS(prob::ODEProblem)
-    state = prob.u0
+    u0 = prob.u0
     eom! = prob.f
 
-    D = length(state); T = eltype(state)
-    du = copy(state)
+    D = length(u0); T = eltype(u0)
+    du = copy(u0)
     J = zeros(T, D, D)
 
-    jeom! = (du, u) -> eom!(0, u, du)
-    jcf = ForwardDiff.JacobianConfig(jeom!, du, state)
-    ForwardDiff_jacob! = (t, u, J) -> ForwardDiff.jacobian!(
+    jeom! = (du, u) -> eom!(du, u, prob.p, 0)
+    jcf = ForwardDiff.JacobianConfig(jeom!, du, u0)
+    ForwardDiff_jacob! = (J, u, p, t) -> ForwardDiff.jacobian!(
     J, jeom!, du, u, jcf)
-    ForwardDiff_jacob!(0, state, J)
+    ForwardDiff_jacob!(J, u0, prob.p, 0)
 
     return ContinuousDS(prob, ForwardDiff_jacob!, J)
 end
 
-function ContinuousDS(state, eom!; tspan=(0.0, 100.0))
+function ContinuousDS(u0, eom!; parameters = nothing, tspan=(0.0, 100.0))
 
-    D = length(state); T = eltype(state)
-    du = copy(state)
+    D = length(u0); T = eltype(u0)
+    du = copy(u0)
     J = zeros(T, D, D)
 
-    problem = ODEProblem{true}(eom!, state, tspan)
+    problem = ODEProblem{true}(eom!, u0, tspan, parameters)
 
-    jeom! = (du, u) -> eom!(0, u, du)
-    jcf = ForwardDiff.JacobianConfig(jeom!, du, state)
-    ForwardDiff_jacob! = (t, u, J) -> ForwardDiff.jacobian!(
+    jeom! = (du, u) -> eom!(du, u, problem.p, 0)
+    jcf = ForwardDiff.JacobianConfig(jeom!, du, u0)
+    ForwardDiff_jacob! = (J, u, p, t) -> ForwardDiff.jacobian!(
     J, jeom!, du, u, jcf)
-    ForwardDiff_jacob!(0, state, J)
+    ForwardDiff_jacob!(J, u0, parameters, 0)
 
     return ContinuousDS(problem, ForwardDiff_jacob!, J)
 end
@@ -146,62 +148,60 @@ dimension(ds::ContinuousDS) = length(ds.prob.u0)
 Base.eltype(ds::ContinuousDS{T,F,J}) where {T, F, J} = T
 state(ds::ContinuousDS) = ds.prob.u0
 
-jacobian(ds::ContinuousDynamicalSystem, t = 0) =
-(ds.jacob!(t, state(ds), ds.J); ds.J)
+jacobian(ds::ContinuousDynamicalSystem) =
+(ds.jacob!(ds.J, state(ds), ds.prob.p, ds.prob.tspan[1]); ds.J)
+
+set_state!(ds::ContinuousDS, u0) = (ds.state .= u0)
 
 #######################################################################################
 #                         Interface to DifferentialEquations                          #
 #######################################################################################
 
-# ODEProblem helper functions
-ODEProblem(ds::ContinuousDS) = ds.prob
-
-ODEProblem(
-ds::ContinuousDS, t::Real, state = ds.prob.u0) =
-ODEProblem{true}(ds.prob.f, state, (zero(t), t),
-callback = ds.prob.callback, mass_matrix = ds.prob.mass_matrix)
-
-ODEProblem(ds::ContinuousDS, tspan::Tuple, state = ds.prob.u0) =
-ODEProblem{true}(ds.prob.f, state, tspan,
-callback = ds.prob.callback, mass_matrix = ds.prob.mass_matrix)
-
 """
-    ODEProblem(ds::ContinuousDS, t, state = ds.prob.u0 [, callback])
-Create a new `ODEProblem` for the given dynamical system that final time `t` (or `tspan`
-for tuple `t`), `state` and if given, also merges the `callback` with other existing
-callbacks currently in `ds.prob`.
+    ODEProblem(ds::ContinuousDS; kwargs...)
+Create a new `ODEProblem` for the given dynamical system by changing specific
+aspects of the existing `ODEProblem`.
+
+Keyword arguments: `state, t, parameters, callback, mass_matrix, tspan`.
+
+If `t` is given, a `tspan` is created with initial time assumed zero. If `tspan`
+is given directly, the keyword `t` is disregarded.
 """
-function ODEProblem(ds::ContinuousDS, t::Real, state, cb)
-    if ds.prob.callback == nothing
-        return ODEProblem{true}(ds.prob.f, state, (zero(t), t),
-        callback = cb, mass_matrix = ds.prob.mass_matrix)
-    else
-        return ODEProblem{true}(ds.prob.f, state, (zero(t), t),
-        callback = CallbackSet(cb, ds.prob.callback),
-        mass_matrix = ds.prob.mass_matrix)
-    end
+function ODEProblem(ds::ContinuousDS;
+                    t = ds.prob.tspan[end],
+                    state = ds.prob.u0,
+                    parameters = ds.prob.p,
+                    callback = ds.prob.callback,
+                    mass_matrix = ds.prob.mass_matrix,
+                    tspan = (zero(t), t))
+
+    return ODEProblem(ds.prob.f, state, tspan, parameters,
+                      mass_matrix = mass_matrix, callback = callback)
 end
 
-"""
-    ODEIntegrator(ds::ContinuousDS, t [, state]; diff_eq_kwargs)
-Return an `ODEIntegrator` to be used directly with the interfaces of
-[`DifferentialEquations.jl`](http://docs.juliadiffeq.org/stable/index.html).
+# Extra callback implementation:
+# """
+#     ODEProblem(ds::ContinuousDS, t, state = ds.prob.u0 [, callback])
+# Create a new `ODEProblem` for the given dynamical system that final time `t` (or `tspan`
+# for tuple `t`), `state` and if given, also merges the `callback` with other existing
+# callbacks currently in `ds.prob`.
+# """
+# function ODEProblem(ds::ContinuousDS, t::Real, state, cb)
+#     if ds.prob.callback == nothing
+#         return ODEProblem{true}(ds.prob.f, state, (zero(t), t),
+#         callback = cb, mass_matrix = ds.prob.mass_matrix)
+#     else
+#         return ODEProblem{true}(ds.prob.f, state, (zero(t), t),
+#         callback = CallbackSet(cb, ds.prob.callback),
+#         mass_matrix = ds.prob.mass_matrix)
+#     end
+# end
 
-`diff_eq_kwargs` is a dictionary `Dict{Symbol, ANY}`
-of keyword arguments
-passed into the `init` of
-[`DifferentialEquations.jl`](http://docs.juliadiffeq.org/stable/index.html),
-for example `Dict(:abstol => 1e-9)`. If you want to specify a solver,
-do so by using the symbol `:solver`, e.g.:
-`Dict(:solver => DP5(), :tstops => 0:0.01:t)`. This requires you to have been first
-`using OrdinaryDiffEq` to access the solvers.
-"""
 function OrdinaryDiffEq.ODEIntegrator(ds::ContinuousDS,
     t, state::Vector = ds.prob.u0; diff_eq_kwargs = DEFAULT_DIFFEQ_KWARGS)
-    prob = ODEProblem(ds, t, state)
+    prob = ODEProblem(ds; t = t, state = state)
     solver, newkw = extract_solver(diff_eq_kwargs)
-    integrator = init(prob, solver; newkw...,
-    save_everystep=false)
+    integrator = init(prob, solver; newkw..., save_everystep=false)
     return integrator
 end
 
@@ -238,17 +238,17 @@ function variational_integrator(ds::ContinuousDS, k::Int, T,
     # dY/dt = J(u) ⋅ Y
     # with J the Jacobian of the vector field at the current state
     # and Y being each of the k deviation vectors
-    veom! = (t, u, du) -> begin
+    veom! = (du, u, p, t) -> begin
         us = view(u, :, 1)
-        f!(t, us, view(du, :, 1))
-        jac!(t, us, J)
+        f!(view(du, :, 1), us, p, t)
+        jac!(J, us, p, t)
         A_mul_B!(view(du, :, 2:k+1), J, view(u, :, 2:k+1))
     end
 
     if typeof(T) <: Real
-        varprob = ODEProblem{true}(veom!, S, (zero(T), T))
+        varprob = ODEProblem{true}(veom!, S, (zero(T), T), ds.prob.p)
     else
-        varprob = ODEProblem{true}(veom!, S, T)
+        varprob = ODEProblem{true}(veom!, S, T, ds.prob.p)
     end
 
     solver, newkw = extract_solver(diff_eq_kwargs)
@@ -282,12 +282,16 @@ const DEFAULT_SOLVER = Vern9()
 # See discrete.jl for the documentation string
 function evolve(ds::ContinuousDS, t = 1.0, state = ds.prob.u0;
     diff_eq_kwargs = DEFAULT_DIFFEQ_KWARGS)
-    prob = ODEProblem(ds, t, state)
+    if typeof(t) <: Real
+        prob = ODEProblem(ds, t = t, state = state)
+    else
+        prob = ODEProblem(ds, tspan = t, state = state)
+    end
     return get_sol(prob, diff_eq_kwargs)[1][end]
 end
 
-evolve!(ds::ContinuousDS, t = 1.0; diff_eq_kwargs = DEFAULT_DIFFEQ_KWARGS) =
-(ds.prob.u0 .= evolve(ds, t, diff_eq_kwargs = diff_eq_kwargs))
+evolve!(ds::ContinuousDS, t; diff_eq_kwargs = DEFAULT_DIFFEQ_KWARGS) =
+(ds.prob.u0 .= evolve(ds, t; diff_eq_kwargs = diff_eq_kwargs); ds.prob.u0)
 
 function extract_solver(diff_eq_kwargs)
     # Extract solver from kwargs
@@ -335,7 +339,7 @@ function use_tstops(prob::ODEProblem)
     if prob.callback == nothing
         return false
     elseif typeof(prob.callback) <: CallbackSet
-        any(x->typeof(x)<:DiscreteCallback, prob.callback.discrete_callbacks)
+        return any(x->typeof(x)<:DiscreteCallback, prob.callback.discrete_callbacks)
     else
         return typeof(prob.callback) <: DiscreteCallback
     end
@@ -353,11 +357,11 @@ function trajectory(ds::ContinuousDS, T;
 
     if typeof(T) <: Real
         t = zero(T):dt:T #time vector
+        prob = ODEProblem(ds; t = T)
     elseif typeof(T) <: Tuple
         t = T[1]:dt:T[2]
+        prob = ODEProblem(ds; tspan = T)
     end
-
-    prob = ODEProblem(ds, T)
 
     return Dataset(get_sol(prob, diff_eq_kwargs, Dict(:saveat => t))[1])
 end